[MDocUpdater] Fix usage of the --delete option
Commit eb7fec62c introduced a breaking change that caused doc removal to
be skipped for members that were removed from a type:

    mdoc: Member Removed: File='/Users/peter/source/android-api-docs/docs/Mono.Android/en/Android.Text/AlteredCharSequence.xml'; Signature='System.Collections.IEnumerator IEnumerable.GetEnumerator ();'
    mdoc: Not deleting 'System.Collections.IEnumerator IEnumerable.GetEnumerator ();'; must be enabled with the --delete option

The error message is now misleading in this case.  I am using the
`--delete` parameter when running the tool, but it no longer works for
removed members.

To help address the scenario fixed in eb7fec62c, a new `--retain-eii`
parameter has been added to mdoc.  When this is parameter is set and
explicit interface implementation docs are not empty, they will not be
removed (even when `--delete` is used).

The `--delete` behavior has otherwise been reverted to its previous
state, and the new `--retain-eii` can be set to opt in to this new EII
specific beahvior.

+++ mdoc/Mono.Documentation/MDocUpdater.cs
@@ -34,6 +34,7 @@ namespace Mono.Documentation
         bool isClassicRun;
         bool verbose;
         bool delete;
+        bool retainEII;
         bool show_exceptions;
         bool no_assembly_versions, ignore_missing_types;
         ExceptionLocations? exceptions;
@@ -188,6 +189,9 @@ namespace Mono.Documentation
             { "delete",
                 "Delete removed members from the XML files.",
                 v => delete = v != null },
+            { "retain-eii",
+                "Retain non-empty explicit interface implementation docs in the XML files instead of regenerating them.",
+                v => retainEII = v != null },
             { "exceptions:",
               "Document potential exceptions that members can generate.  {SOURCES} " +
                 "is a comma-separated list of:\n" +
@@ -1592,6 +1596,10 @@ namespace Mono.Documentation
                     if (!no_assembly_versions && UpdateAssemblyVersions (oldmember, type.Module.Assembly, new string[] { GetAssemblyVersion (type.Module.Assembly) }, false))
                         continue;
 
+                    // If a private EII has existing docs we may not want to delete and regenerate them
+                    if (isprivateeii && retainEII && MemberDocsHaveUserContent (oldmember))
+                        continue;
+
                     DeleteMember ("Member Removed", output, oldmember, todelete, type);
                     statisticsCollector.AddMetric(typeEntry.Framework.Name, StatisticsItem.Members, StatisticsMetrics.Removed);
                     continue;
@@ -1939,11 +1947,10 @@ namespace Mono.Documentation
                     signature);
 
             // Identify all of the different states that could affect our decision to delete the member
-            bool duplicated = reason.Contains("Duplicate Member");
             bool shouldPreserve = !string.IsNullOrWhiteSpace (PreserveTag);
+            // For historical consistency, docs with no content will be removed even if the --delete param is not used.
             bool hasContent = MemberDocsHaveUserContent (member);
-            //When the member is NOT PRESERVED, the member has NO CONTENT or  is DUPLICATED, then it should be deleted
-            bool shouldDelete = !shouldPreserve && (delete && (!hasContent || duplicated));
+            bool shouldDelete = !shouldPreserve && (delete || !hasContent);
 
             bool unifiedRun = HasDroppedNamespace (type);
